pyproject.toml
[project]
name = "riksbank-app"
version = "0.1.0"
description = "riskbank test task"
authors = [
    {name = "Reiny24",email = "dima.guk2004@gmail.com"}
]
readme = "README.md"
requires-python = "3.10.8"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "schedule (>=1.2.2,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

